Added Skeleton of Rating system and Restructured Components.
I took the original ./pages/search.js file and broke it down into several components which now live in the ./pages/search-page folder.

Linked Index.js to the new Search.js in this folder.

Added skeleton of rating system.
