+++ models/speedSchema.js
@@ -58,6 +58,10 @@ const SEPracticeSchema = new mongoose.Schema({
     enum: ['Student', 'Practitioner'], // enum used to limit to specified values
     required: true,
   },
+  userRating: {
+    type: Number,
+    required: false
+  }
 });
 
 export default mongoose.models.SEPractice || mongoose.model('SEPractice', SEPracticeSchema);
