+++ pages/index.js
@@ -29,7 +29,7 @@ export default function Home() {
         <button type="submit">Submit</button>
       </form>
       {/* Link to search page */}
-      <Link href="/search">
+      <Link href="/search-page/Search">
         Go to Search page
       </Link>
     </div>
